refactor: AuthService to AuthCommandHandler and DI with interface
