+++ src/main/kotlin/com/template/auth/service/AuthCommandHandler.kt
@@ -1,0 +1,6 @@
+package com.template.auth.service
+
+interface AuthCommandHandler {
+    fun updateAccessToken(refreshToken: String): String
+    fun login(email: String, password: String): Pair<String, String>
+}
