+++ src/main/kotlin/com/template/auth/service/impl/AuthCommandHandlerImpl.kt
@@ -1,7 +1,8 @@
-package com.template.auth.service
+package com.template.auth.service.impl
 
 import com.template.auth.exception.AuthenticateException
 import com.template.auth.exception.LoginException
+import com.template.auth.service.AuthCommandHandler
 import com.template.auth.tools.JwtTokenUtil
 import com.template.user.domain.UserRepository
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
@@ -9,14 +10,13 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
-class AuthService(
+class AuthCommandHandlerImpl(
     private val jwtTokenUtil: JwtTokenUtil,
     private val userRepository: UserRepository,
     private val encoder: BCryptPasswordEncoder
-) {
-
+) : AuthCommandHandler {
     @Transactional(readOnly = true)
-    fun updateAccessToken(refreshToken: String): String {
+    override fun updateAccessToken(refreshToken: String): String {
         if (!jwtTokenUtil.isTokenExpired(refreshToken)) {
             val userId = jwtTokenUtil.extractUserId(refreshToken)
             if (userRepository.existsById(userId)) {
@@ -26,7 +26,7 @@ class AuthService(
     }
 
     @Transactional(readOnly = true)
-    fun login(email: String, password: String): Pair<String, String> {
+    override fun login(email: String, password: String): Pair<String, String> {
         val user = userRepository.findByEmail(email).orElseThrow { LoginException() }
         if (!encoder.matches(password, user.password)) throw LoginException()
         val accessToken = jwtTokenUtil.generateAccessToken(user.id!!)
