+++ src/main/kotlin/com/template/user/service/UserQueryHandler.kt
@@ -1,0 +1,3 @@
+package com.template.user.service
+
+class UserQueryHandler
