--- src/main/kotlin/com/template/user/service/UserService.kt
@@ -1,13 +1,0 @@
-package com.template.user.service
-
-import com.template.user.domain.UserRepository
-import com.template.user.service.dto.UserDto
-import org.springframework.stereotype.Service
-
-@Service
-class UserService(
-    private val userRepository: UserRepository
-) {
-
-    fun getUserInfo(userDto: UserDto) = userDto
-}
