+++ src/main/kotlin/com/template/user/service/impl/UserQueryHandlerImpl.kt
@@ -1,0 +1,3 @@
+package com.template.user.service.impl
+
+class UserQueryHandlerImpl
